requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.5
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pyimgur==0.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
schemdraw==0.14
scipy==1.7.3
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
wincertstore==0.2
